added code coverage and renamed test_utils to testing_utils
