+++ .coveragerc
@@ -1,0 +1,11 @@
+[pytest]
+addopts = --cov --cov-report=term
+testpaths = sentinelhub
+
+[run]
+source = sentinelhub
+
+[report]
+omit =
+    config.json
+    .utmzones.geojson
