+++ .travis.yml
@@ -24,11 +24,14 @@ script:
 - radon cc sentinelhub/*.py -a -nb
 - >
   if [[ ($TRAVIS_PULL_REQUEST == true) || ( $TRAVIS_PYTHON_VERSION != '3.7') ]] ;
-  then pytest --cov ./tests/ \
+  then pytest --cov \
     --ignore tests/test_ogc.py \
     --ignore tests/test_data_request.py \
     --ignore tests/test_aws.py \
     --ignore tests/test_aws_safe.py \
     --ignore tests/test_commands.py ;
-  else pytest --cov ./tests/ ;
+  else pytest --cov ;
   fi
+
+after_success:
+- if [ $TRAVIS_PYTHON_VERSION == '3.7' ] && [ "$TRAVIS_PULL_REQUEST" = "false" ]; then codecov; fi
