--- docs/source/test_utils.rst
@@ -1,6 +1,0 @@
-test_utils
-==========
-
-.. automodule:: sentinelhub.test_utils
-    :members:
-    :show-inheritance:
