+++ docs/source/testing_utils.rst
@@ -1,0 +1,6 @@
+testing_utils
+=============
+
+.. automodule:: sentinelhub.testing_utils
+    :members:
+    :show-inheritance:
