+++ requirements-dev.txt
@@ -1,5 +1,6 @@
 pytest>=4.0.0
 pytest-cov
+codecov
 pylint>=2.2.0
 twine
 radon
