+++ sentinelhub/__init__.py
@@ -30,6 +30,6 @@ from .geo_utils import bbox_to_dimensions, bbox_to_resolution, get_image_dimensi
     wgs84_to_utm, to_wgs84, utm_to_pixel, pixel_to_utm, wgs84_to_pixel, get_utm_crs, transform_point, transform_bbox
 from .time_utils import next_date, prev_date, get_current_date
 
-from .test_utils import TestSentinelHub, TestCaseContainer
+from .testing_utils import TestSentinelHub, TestCaseContainer
 
 from ._version import __version__
